refactor(server): extract publishing-queue routes (1/3 of publishing)
First of the 3-way publishing split (queue / channels / publish). Moves
the queue CRUD + lifecycle routes into src/server/routes/publishing-queue.js,
mounted `app.use('/api/publishing', publishingQueueRouter)` (channels +
the publish dispatcher follow in their own files, all sharing the
/api/publishing mount — guard-safe since separate files).

14 routes moved (sync, log, republish, unpublish, queue list/CRUD,
queue/:id archive/unarchive/title/reset-channel/request-review,
backfill-queue). MIXED auth -> per-route.

INTENTIONAL route change: POST /api/publishing/backfill-queue was
registered 3x identically (only the first reachable in Express; the other
2 dead/shadowed). Dropped the 2 dead dupes — route count 213 -> 211,
snapshot regenerated.

Deps all already extracted: pool, requireAuth/verifyBrandAccess (auth),
buildXOAuthHeader (x — sync/republish poll X directly), createHmac/
randomBytes (crypto, OAuth1 signing), RESEND_API_KEY (request-review
emails a reviewer; module-level const off process.env).

Test plan:
- node --check server.js + module -> OK
- npm run lint (no-undef) -> clean (caught buildXOAuthHeader/createHmac/
  randomBytes/RESEND_API_KEY; added)
- module boot-load (import()) -> resolves
- npm run routes:snapshot -> 211 (the intentional dupe removal)
- route-inventory guard -> 211 byte-identical (20 publishing paths: 14
  queue reconstructed + 4 channels + 2 publish still inline)
- vitest -> 93/93

Rollback: revert; restores inline queue routes incl. the 2 dead dupes +
the 213 snapshot.
